moved DecoderType to ui package, changed comments to start with Uppercase, added setter for startTimeMs
