+++ app/src/main/java/org/jellyfin/mobile/player/ui/DecoderType.kt
@@ -1,4 +1,4 @@
-package org.jellyfin.mobile.player.interaction
+package org.jellyfin.mobile.player.ui
 
 /**
  * Represents the type of decoder
