+++ app/src/main/java/org/jellyfin/mobile/player/ui/PlayerFragment.kt
@@ -33,7 +33,6 @@ import org.jellyfin.mobile.databinding.ExoPlayerControlViewBinding
 import org.jellyfin.mobile.databinding.FragmentPlayerBinding
 import org.jellyfin.mobile.player.PlayerException
 import org.jellyfin.mobile.player.PlayerViewModel
-import org.jellyfin.mobile.player.interaction.DecoderType
 import org.jellyfin.mobile.player.interaction.PlayOptions
 import org.jellyfin.mobile.utils.Constants
 import org.jellyfin.mobile.utils.Constants.DEFAULT_CONTROLS_TIMEOUT_MS
